# buffer: raw_combined allocations buffer and ref count together