--- a/src/include/buffer.h
+++ b/src/include/buffer.h
@@ -386,6 +386,10 @@ namespace buffer CEPH_BUFFER_API {
       return *this;
     }
 
+    unsigned get_num_buffers() const { return _buffers.size(); }
+    const ptr& front() const { return _buffers.front(); }
+    const ptr& back() const { return _buffers.back(); }
+
     unsigned get_memcopy_count() const {return _memcopy_count; }
     const std::list<ptr>& buffers() const { return _buffers; }
     void swap(list& other);