--- a/src/rbd_replay/BufferReader.cc
+++ b/src/rbd_replay/BufferReader.cc
@@ -16,9 +16,9 @@ BufferReader::BufferReader(int fd, size_t min_bytes, size_t max_bytes)
 int BufferReader::fetch(bufferlist::iterator **it) {
   if (m_bl_it.get_remaining() < m_min_bytes) {
     ssize_t bytes_to_read = ROUND_UP_TO(m_max_bytes - m_bl_it.get_remaining(),
-                                        CEPH_BUFFER_APPEND_SIZE);
+                                        CEPH_PAGE_SIZE);
     while (!m_eof_reached && bytes_to_read > 0) {
-      int r = m_bl.read_fd(m_fd, CEPH_BUFFER_APPEND_SIZE);
+      int r = m_bl.read_fd(m_fd, CEPH_PAGE_SIZE);
       if (r < 0) {
         return r;
       }