--- a/src/test/bufferlist.cc
+++ b/src/test/bufferlist.cc
@@ -457,7 +457,7 @@ TEST(BufferPtr, constructors) {
     bufferptr ptr(std::move(original));
     EXPECT_TRUE(ptr.have_raw());
     EXPECT_FALSE(original.have_raw());
-    EXPECT_EQ(str.compare(0, str.size(), ptr.c_str()), 0);
+    EXPECT_EQ(str.compare(0, str.size(), ptr.c_str(), ptr.length()), 0);
     EXPECT_EQ(1, ptr.raw_nref());
   }
 }