@branch-predictor 2016-02-15
![write-1024](https://cloud.githubusercontent.com/assets/12440680/13042231/dfff92d2-d3be-11e5-90ef-7666aa770e0c.png)
![write-4096](https://cloud.githubusercontent.com/assets/12440680/13042232/e001af86-d3be-11e5-8373-0146c628331e.png)
![write-65536](https://cloud.githubusercontent.com/assets/12440680/13042233/e00232c6-d3be-11e5-8259-78fa88bd7f60.png)

This is how rados bench 120 write -t 128 went with and without patch. This patch improved performance on small IO, but the difference is smaller due to more disk flushing occurring more frequently. For large I/O, the difference is close to none.
I assume that with bluestore the performance difference will be even greater than that.
(Complete data here: http://ceph.predictor.org.pl/bufferlist_patch.xlsx, raw data at http://ceph.predictor.org.pl/chunktest_org.tar.gz and http://ceph.predictor.org.pl/chunktest_patch.tar.gz)